get all projects for given identity in identity module
